修复由于日志模块更新导致的问题

+++ result/builder.go
@@ -23,7 +23,7 @@ func (b *builder) Success() *builder {
 }
 
 func (b *builder) ErrorCode(code int) *builder {
-	b.result.Code = code
+	b.result.Code = int64(code)
 	return b
 }
